move platform_adaptive into common
