+++ lib/common/platform_adaptive.dart
